+++ backend/database/migrations/2024_05_17_234245_employees.php
@@ -11,7 +11,7 @@ return new class extends Migration
     {
         Schema::create($this->table, function (Blueprint $table) {
             $table->uuid('id')->primary();
-            $table->bigInteger('sequence');
+            $table->integer('sequence');
             $table->string('name', 100);
             $table->string('email', 100);
             $table->boolean('deleted')->default(0);
